requirements.txt
azure-search-documents
azure-identity
langchain_community
langchain_openai

python-dotenv